# docs/swagger.yaml
basePath: /
definitions:
  models.CounterResponse:
    properties:
      count:
        type: integer
      key:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
  models.FullNameRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    type: object
  models.FullNameResponse:
    properties:
      name:
        type: string
    type: object
  models.MoxyResponse:
    properties:
      data:
        type: string
    type: object
info:
  contact:
    email: shubham.sinha@angelbroking.com
    name: Shubham Sinha
  description: Go Example Project
  termsOfService: https://swagger.io/terms/
  title: Go Example Project
  version: "1.0"
paths:
  /counter/create:
    post:
      description: Creates a new counter
      operationId: createCounter
      parameters:
      - description: counter key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Creates a new counter
      tags:
      - counter
  /counter/current:
    get:
      description: Get the current value of counter
      operationId: currentCount
      parameters:
      - description: counter key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CounterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get the current value of counter
      tags:
      - counter
  /counter/decrement:
    put:
      description: Decrement an existing counter
      operationId: decrementCounter
      parameters:
      - description: counter key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Decrement an existing counter
      tags:
      - counter
  /counter/increment:
    put:
      description: Increment an existing counter
      operationId: incrementCounter
      parameters:
      - description: counter key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Increment an existing counter
      tags:
      - counter
  /fullName:
    post:
      consumes:
      - application/json
      description: Gets the full name from the first name and last name
      operationId: fullName
      parameters:
      - description: first name and last name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FullNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FullNameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Gets the full name from the first name and last name
      tags:
      - fullName
  /moxy:
    get:
      description: Get the moxy response
      operationId: moxy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MoxyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get the moxy response
      tags:
      - moxy
swagger: "2.0"